apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
httpx
qdrant-client
opensearch-py
boto3
redis
arq
python-multipart
tree-sitter
tree-sitter-languages
PyMuPDF
pytesseract
Pillow
tenacity
structlog
